Move unstable tail detection to snapshot
